+++ packages/stryker/test/unit/SandboxSpec.ts
@@ -20,6 +20,9 @@ import * as fileUtils from '../../src/utils/fileUtils';
 import currentLogMock from '../helpers/logMock';
 import TestRunnerDecorator from '../../src/isolated-runner/TestRunnerDecorator';
 
+const OVERHEAD_TIME_MS = 0;
+const SANDBOX_INDEX = 3;
+
 describe('Sandbox', () => {
   let options: Config;
   let files: File[];
@@ -66,18 +69,18 @@ describe('Sandbox', () => {
   describe('create()', () => {
 
     it('should copy input files when created', async () => {
-      await Sandbox.create(options, 3, files, null);
+      await Sandbox.create(options, SANDBOX_INDEX, files, null, OVERHEAD_TIME_MS);
       expect(fileUtils.writeFile).calledWith(expectedTargetFileToMutate, files[0].content);
       expect(fileUtils.writeFile).calledWith(path.join(sandboxFolder, 'file2'), files[1].content);
     });
 
     it('should copy a local file when created', async () => {
-      await Sandbox.create(options, 3, [new File('localFile.js', 'foobar')], null);
+      await Sandbox.create(options, SANDBOX_INDEX, [new File('localFile.js', 'foobar')], null, OVERHEAD_TIME_MS);
       expect(fileUtils.writeFile).calledWith(path.join(sandboxFolder, 'localFile.js'), Buffer.from('foobar'));
     });
 
     it('should have created the isolated test runner', async () => {
-      await Sandbox.create(options, 3, files, null);
+      await Sandbox.create(options, SANDBOX_INDEX, files, null, OVERHEAD_TIME_MS);
       const expectedSettings: IsolatedRunnerOptions = {
         port: 46,
         strykerOptions: options,
@@ -88,19 +91,19 @@ describe('Sandbox', () => {
     });
 
     it('should have created a sandbox folder', async () => {
-      await Sandbox.create(options, 3, files, testFrameworkStub);
+      await Sandbox.create(options, SANDBOX_INDEX, files, testFrameworkStub, OVERHEAD_TIME_MS);
       expect(TempFolder.instance().createRandomFolder).to.have.been.calledWith('sandbox');
     });
 
     it('should symlink node modules in sandbox directory if exists', async () => {
-      await Sandbox.create(options, 3, files, testFrameworkStub);
+      await Sandbox.create(options, SANDBOX_INDEX, files, testFrameworkStub, OVERHEAD_TIME_MS);
       expect(findNodeModulesStub).calledWith(process.cwd());
       expect(symlinkJunctionStub).calledWith('node_modules', path.join(sandboxFolder, 'node_modules'));
     });
 
     it('should not symlink node modules in sandbox directory if no node_modules exist', async () => {
       findNodeModulesStub.resolves(null);
-      await Sandbox.create(options, 3, files, testFrameworkStub);
+      await Sandbox.create(options, SANDBOX_INDEX, files, testFrameworkStub, OVERHEAD_TIME_MS);
       expect(log.warn).calledWithMatch('Could not find a node_modules');
       expect(log.warn).calledWithMatch(process.cwd());
       expect(symlinkJunctionStub).not.called;
@@ -109,7 +112,7 @@ describe('Sandbox', () => {
     it('should log a warning if "node_modules" already exists in the working folder', async () => {
       findNodeModulesStub.resolves('node_modules');
       symlinkJunctionStub.rejects(createFileAlreadyExistsError());
-      await Sandbox.create(options, 3, files, testFrameworkStub);
+      await Sandbox.create(options, SANDBOX_INDEX, files, testFrameworkStub, OVERHEAD_TIME_MS);
       expect(log.warn).calledWithMatch(normalizeWhiteSpaces(
         `Could not symlink "node_modules" in sandbox directory, it is already created in the sandbox. 
         Please remove the node_modules from your sandbox files. Alternatively, set \`symlinkNodeModules\` 
@@ -117,17 +120,17 @@ describe('Sandbox', () => {
     });
 
     it('should log a warning if linking "node_modules" results in an unknown error', async () => {
-      findNodeModulesStub.resolves('basepath/node_modules');
+      findNodeModulesStub.resolves('basePath/node_modules');
       const error = new Error('unknown');
       symlinkJunctionStub.rejects(error);
-      await Sandbox.create(options, 3, files, testFrameworkStub);
+      await Sandbox.create(options, SANDBOX_INDEX, files, testFrameworkStub, OVERHEAD_TIME_MS);
       expect(log.warn).calledWithMatch(normalizeWhiteSpaces(
-        `Unexpected error while trying to symlink "basepath/node_modules" in sandbox directory.`), error);
+        `Unexpected error while trying to symlink "basePath/node_modules" in sandbox directory.`), error);
     });
 
     it('should symlink node modules in sandbox directory if `symlinkNodeModules` is `false`', async () => {
       options.symlinkNodeModules = false;
-      await Sandbox.create(options, 3, files, testFrameworkStub);
+      await Sandbox.create(options, SANDBOX_INDEX, files, testFrameworkStub, OVERHEAD_TIME_MS);
       expect(symlinkJunctionStub).not.called;
       expect(findNodeModulesStub).not.called;
     });
@@ -135,7 +138,7 @@ describe('Sandbox', () => {
 
   describe('run()', () => {
     it('should run the testRunner', async () => {
-      const sut = await Sandbox.create(options, 3, files, null);
+      const sut = await Sandbox.create(options, SANDBOX_INDEX, files, null, 0);
       await sut.run(231313, 'hooks');
       expect(testRunner.run).to.have.been.calledWith({
         timeout: 231313,
@@ -163,21 +166,21 @@ describe('Sandbox', () => {
     });
 
     it('should save the mutant to disk', async () => {
-      const sut = await Sandbox.create(options, 3, files, null);
+      const sut = await Sandbox.create(options, SANDBOX_INDEX, files, null, OVERHEAD_TIME_MS);
       await sut.runMutant(transpiledMutant);
       expect(fileUtils.writeFile).calledWith(expectedTargetFileToMutate, Buffer.from('mutated code'));
       expect(log.warn).not.called;
     });
 
     it('should nog log a warning if test selection was failed but already reported', async () => {
-      const sut = await Sandbox.create(options, 3, files, null);
+      const sut = await Sandbox.create(options, SANDBOX_INDEX, files, null, OVERHEAD_TIME_MS);
       transpiledMutant.mutant.testSelectionResult = TestSelectionResult.FailedButAlreadyReported;
       await sut.runMutant(transpiledMutant);
       expect(log.warn).not.called;
     });
 
     it('should log a warning if tests could not have been selected', async () => {
-      const sut = await Sandbox.create(options, 3, files, null);
+      const sut = await Sandbox.create(options, SANDBOX_INDEX, files, null, OVERHEAD_TIME_MS);
       transpiledMutant.mutant.testSelectionResult = TestSelectionResult.Failed;
       await sut.runMutant(transpiledMutant);
       const expectedLogMessage = `Failed find coverage data for this mutant, running all tests. This might have an impact on performance: ${transpiledMutant.mutant.toString()}`;
@@ -185,20 +188,23 @@ describe('Sandbox', () => {
     });
 
     it('should filter the scoped tests', async () => {
-      const sut = await Sandbox.create(options, 3, files, testFrameworkStub);
+      const sut = await Sandbox.create(options, SANDBOX_INDEX, files, testFrameworkStub, OVERHEAD_TIME_MS);
       await sut.runMutant(transpiledMutant);
       expect(testFrameworkStub.filter).to.have.been.calledWith(transpiledMutant.mutant.selectedTests);
     });
 
     it('should provide the filter code as testHooks and correct timeout', async () => {
-      const sut = await Sandbox.create(options, 3, files, testFrameworkStub);
+      options.timeoutMs = 1000;
+      const overheadTimeMS = 42;
+      const totalTimeSpend = 12;
+      const sut = await Sandbox.create(options, SANDBOX_INDEX, files, testFrameworkStub, overheadTimeMS);
       await sut.runMutant(transpiledMutant);
-      const expectedRunOptions = { testHooks: wrapInClosure(testFilterCodeFragment), timeout: 12 * 23 + 1000 };
+      const expectedRunOptions = { testHooks: wrapInClosure(testFilterCodeFragment), timeout: totalTimeSpend * options.timeoutFactor + options.timeoutMs + overheadTimeMS };
       expect(testRunner.run).calledWith(expectedRunOptions);
     });
 
     it('should have reset the source file', async () => {
-      const sut = await Sandbox.create(options, 3, files, null);
+      const sut = await Sandbox.create(options, SANDBOX_INDEX, files, null, OVERHEAD_TIME_MS);
       await sut.runMutant(transpiledMutant);
       let timesCalled = writeFileStub.getCalls().length - 1;
       let lastCall = writeFileStub.getCall(timesCalled);
@@ -206,7 +212,7 @@ describe('Sandbox', () => {
     });
 
     it('should not filter any tests when testFramework = null', async () => {
-      const sut = await Sandbox.create(options, 3, files, null);
+      const sut = await Sandbox.create(options, SANDBOX_INDEX, files, null, OVERHEAD_TIME_MS);
       const mutant = new TestableMutant('2', createMutant(), new SourceFile(new File('', '')));
       sut.runMutant(new TranspiledMutant(mutant, { outputFiles: [new File(expectedTargetFileToMutate, '')], error: null }, true));
       expect(fileUtils.writeFile).not.calledWith(expectedTestFrameworkHooksFile);